requirements.txt
datasets
transformers
torch
trl
matplotlib
pyyaml
pytest